Recreate empty folder, needed for testcases

+++ test/de/uni_hildesheim/sse/kernel_haven/undertaker/UndertakerWrapperTest.java
@@ -43,7 +43,9 @@ public class UndertakerWrapperTest {
     @AfterClass
     public static void afterClass() {
         for (File file : RESOURCE_DIR.listFiles()) {
-            file.delete();
+            if (!file.getName().equals(".gitignore")) {
+                file.delete();
+            }
         }
     }
